Fix issues and rewrite loopgen(NaryExpr)

+++ test/test_base.cpp
@@ -155,42 +155,42 @@ void imax_test()
 {
     select_test<int32_t, int32_t>(
         [] (Expr s) { return _max(s, _i32(10)); },
-        [] (int32_t s) { return std::max(s, static_cast<int32_t>(10)); });
+        [] (int32_t s) { return std::max(s, 10); });
 }
 
 void umax_test()
 {
     select_test<uint32_t, uint32_t>(
         [] (Expr s) { return _max(s, _u32(10)); },
-        [] (uint32_t s) { return static_cast<uint32_t>(std::max(s, static_cast<uint32_t>(10))); });
+        [] (uint32_t s) { return static_cast<uint32_t>(std::max(s, 10u)); });
 }
 
 void fmax_test()
 {
     select_test<float, float>(
         [] (Expr s) { return _max(s, _f32(10)); },
-        [] (float s) { return static_cast<float>(std::max(s, static_cast<float>(10))); });
+        [] (float s) { return static_cast<float>(std::max(s, 10.0f)); });
 }
 
 void imin_test()
 {
     select_test<int32_t, int32_t>(
         [] (Expr s) { return _min(s, _i32(10)); },
-        [] (int32_t s) { return std::min(s, static_cast<int32_t>(10)); });
+        [] (int32_t s) { return std::min(s, 10); });
 }
 
 void umin_test()
 {
     select_test<uint32_t, uint32_t>(
         [] (Expr s) { return _min(s, _u32(10)); },
-        [] (uint32_t s) { return static_cast<uint32_t>(std::min(s, static_cast<uint32_t>(10))); });
+        [] (uint32_t s) { return static_cast<uint32_t>(std::min(s, 10u)); });
 }
 
 void fmin_test()
 {
     select_test<float, float>(
         [] (Expr s) { return _min(s, _f32(10)); },
-        [] (float s) { return static_cast<float>(std::min(s, static_cast<float>(10))); });
+        [] (float s) { return static_cast<float>(std::min(s, 10.0f)); });
 }
 
 void fsqrt_test()
